fixing bug of all component
