+++ app/(root)/(client)/orderHistory/page.tsx
@@ -89,7 +89,7 @@ const OrderHistoryPage = () => {
         };
 
         fetchOrders();
-    }, []);
+    }, [currentPage]);
 
     const formatVND = (amount: number) => {
         return new Intl.NumberFormat('vi-VN', {
